+++ features/auth/client/src/commonMain/kotlin/configurators/BearerAuthHttpClientConfigurator.kt
@@ -1,6 +1,5 @@
 package dev.inmo.wishlist.features.auth.client.configurators
 
-import dev.inmo.micro_utils.coroutines.MutableRedeliverStateFlow
 import dev.inmo.micro_utils.coroutines.runCatchingLogging
 import io.ktor.client.HttpClientConfig
 import io.ktor.client.call.body
@@ -14,13 +13,7 @@ import io.ktor.http.ContentType
 import io.ktor.http.contentType
 import io.ktor.http.encodedPath
 import io.ktor.http.isSuccess
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.dropWhile
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.mapNotNull
 import dev.inmo.wishlist.features.auth.client.AuthCredentialsStorage
-import dev.inmo.wishlist.features.auth.client.ClientAuthFeature
 import dev.inmo.wishlist.features.auth.common.Constants
 import dev.inmo.wishlist.features.auth.common.models.AuthCredentials
 import dev.inmo.wishlist.features.auth.common.models.RefreshRequest
@@ -32,11 +25,15 @@ class BearerAuthHttpClientConfigurator(
 ) : HttpClientConfigurator {
     private val refreshPath = "${Constants.prefixPathPart}/${Constants.refreshPathPart}"
     private val loginPath = "${Constants.prefixPathPart}/${Constants.loginPathPart}"
+    private val registerPath = "${Constants.prefixPathPart}/${Constants.registerPathPart}"
+    private val isRegistrationAvailablePath = "${Constants.prefixPathPart}/${Constants.isRegistrationAvailablePathPart}"
 
     override fun HttpClientConfig<*>.configure() {
         val storage = storage
         val refreshPath = refreshPath
         val loginPath = loginPath
+        val registerPath = registerPath
+        val configPath = isRegistrationAvailablePath
         install(Auth) {
             bearer {
                 cacheTokens = false
@@ -66,7 +63,10 @@ class BearerAuthHttpClientConfigurator(
                 }
                 sendWithoutRequest { request ->
                     val path = request.url.encodedPath
-                    !path.endsWith(loginPath) && !path.endsWith(refreshPath)
+                    !path.endsWith(loginPath) &&
+                        !path.endsWith(refreshPath) &&
+                        !path.endsWith(registerPath) &&
+                        !path.endsWith(configPath)
                 }
             }
         }