+++ features/auth/common/src/commonMain/kotlin/Constants.kt
@@ -6,4 +6,6 @@ object Constants {
     const val refreshPathPart = "refresh"
     const val logoutPathPart = "logout"
     const val getMePathPart = "getMe"
+    const val registerPathPart = "register"
+    const val isRegistrationAvailablePathPart = "is_registration_available"
 }