+++ features/auth/server/src/commonMain/kotlin/Config.kt
@@ -8,5 +8,6 @@ import kotlin.time.Duration.Companion.minutes
 @Serializable
 data class Config(
     val tokenTtl: Duration = 15.minutes,
-    val refreshTokenTtl: Duration = 7.days
+    val refreshTokenTtl: Duration = 7.days,
+    val enableRegistration: Boolean = false,
 )